projects/diabetes_risk/requirements.txt
marimo
matplotlib
numpy
pandas
python-dotenv
scikit_learn
xgboost